requirement.txt
amqp==2.5.0
anyjson==0.3.3
apipkg==1.5
asn1crypto==0.22.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.5.1
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.8.0
billiard==3.6.0.0
boto==2.48.0
cachetools==3.0.0
celery==4.3.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0.3
cycler==0.10.0
decorator==4.1.2
defusedxml==0.6.0
Django==1.11
django-appconf==1.0.3
django-axes==5.0.5
django-bootstrap-form==3.4
django-bulk-update==2.2.0
django-celery==3.2.1
django-cleanup==4.0.0
django-cors-headers==2.1.0
django-defender==0.6.1
amqp==2.5.0
anyjson==0.3.3
apipkg==1.5
asn1crypto==0.22.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.5.1
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.8.0
billiard==3.6.0.0
boto==2.48.0
cachetools==3.0.0
celery==4.3.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0.3
cycler==0.10.0
decorator==4.1.2
defusedxml==0.6.0
Django==1.11
django-appconf==1.0.3
django-axes==5.0.5
django-bootstrap-form==3.4
django-bulk-update==2.2.0
django-celery==3.2.1
django-cleanup==4.0.0
django-cors-headers==2.1.0
django-defender==0.6.1
django-helpdesk==0.2.17
django-ipware==2.1.0
django-markdown-deux==1.0.5
django-redis==4.8.0
django-redis-cache==1.7.1
django-ses==0.8.2
django-tinymce==2.8.0
djangorestframework==3.6.3
elasticsearch==6.3.0
elasticsearch-dsl==6.2.1
email-reply-parser==0.5.9
enum34==1.1.6
execnet==1.6.0
Flask==0.12.2
flower==0.9.2
futures==3.1.1
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
gunicorn==19.7.1
hashids==1.2.0
httplib2==0.10.3
idna==2.6
ipaddress==1.0.18
ipdb==0.10.3
ipython==5.4.1
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.9.6
joblib==0.13.2
kiwisolver==1.1.0
kombu==4.6.3
lxml==4.4.1
markdown2==2.3.8
MarkupSafe==1.0
matplotlib==3.1.0
mongoengine==0.10.7
more-itertools==7.0.0
mysqlclient==1.3.10
numpy==1.16.3
oauth2client==4.1.3
oauthlib==3.0.1
olefile==0.44
pandas==0.24.2
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
pluggy==0.11.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.8.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pycrypto==2.6.1
pyfcm==1.4.5
Pygments==2.2.0
PyJWT==1.5.2
pymongo==3.7.2
pyOpenSSL==17.2.0
pyparsing==2.4.0
PySocks==1.6.8
pytest==4.5.0
pytest-forked==1.0.2
pytest-xdist==1.28.0
python-dateutil==2.6.1
python-http-client==3.0.0
python-logstash==0.4.6
python-redis-lock==3.2.0
python-social-auth==0.2.1
python3-openid==3.1.0
pytz==2017.2
redis==3.2.0
requests==2.18.4
requests-oauthlib==0.3.1
requests-toolbelt==0.8.0
rsa==4.0
scandir==1.5
scikit-learn==0.21.2
scipy==1.3.0
sendgrid==5.3.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.10.0
soupsieve==1.9.3
tornado==4.5.2
traitlets==4.3.2
twilio==4.6.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
views==0.3
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.12.2
xlrd==1.2.0
